faille csrf et variable twig

+++ app/src/Controller/AdminController.php
@@ -17,9 +17,15 @@ final class AdminController extends AbstractController
     }
 
     #[Route('/admin/delete/{id<[0-9]+>}', name: 'app_admin_delete')]
-    public function delete(Item $item): Response
+    public function delete(Item $item, Request $request): Response
     {
-        $id = $item->getId();
+        $submittedToken = $request->getPayload()->get('token');
+
+        if (!$this->isCsrfTokenValid($_ENV['CSRF_SECRET'] . $item->getId(), $submittedToken)) {
+            $this->addFlash('danger', 'Token invalide');
+
+            return $this->redirectToRoute('app_item_show', ['id' => $item->getId()]);
+        }
 
         $this->em->remove($item);
         $this->em->flush();
